# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CERBERUS Commerce OS — API Sözleşmesi (T3.5)
  version: "3.0-faz3"
  description: >
    Tüm uçlar `cerberus_session` imzalı JWT çerezi ile korunur (açık uçlar: /api/auth/login, /api/health).
    Kimliksiz istek 401; rol/storeCode ihlali 403; doğrulama hatası 422; sunucu hatası 500 + correlationId döner.
servers:
  - url: /
...
paths:
...
  /api/orders:
    get:
      tags: [orders]
      summary: Sipariş listesi + SQL KPI + sayfalama
      parameters:
        - { name: storeCode, in: query, schema: { type: string } , description: "STORE_USER için yok sayılır (F-11)" }
        - { name: cargoStatus, in: query, schema: { type: string } }
        - { name: pshBatchNo, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 200, maximum: 2000 } }
      responses:
        "200": { description: "orders, stores, batches, auditLogs, users(güvenli), kpis, pagination" }
        "401": { description: Oturum gerekli }
    post:
      tags: [orders]
      summary: Yeni sipariş (40-kolon XLS sözleşmesi; orderCreateSchema)
      responses:
        "200": { description: "message + order" }
        "401": { description: Oturum gerekli }
        "422": { description: Doğrulama hatası }
...