# src/app/api/orders/route.ts
import { NextResponse } from "next/server";
import { db } from "@/db";
import { orders, stores, pshBatches, auditLogs, users } from "@/db/schema";
import { requireUser, isDenied, resolveStoreScope } from "@/lib/guards";
import { parseBody, orderCreateSchema } from "@/lib/validation";
import { handleRouteError } from "@/lib/apiResponse";
import { desc, eq, and, inArray, count, sql } from "drizzle-orm";

export async function GET(req: Request) {
  try {
    // Kimlik doğrulama zorunlu (F-02/F-05): anonim istek 401 alır.
    const gate = await requireUser();
    if (isDenied(gate)) return gate.response;
    const currentUser = gate.user;


    const { searchParams } = new URL(req.url);
    const requestedStore = searchParams.get("storeCode") || "ALL";
    const cargoStatus = searchParams.get("cargoStatus");
    const pshBatchNo = searchParams.get("pshBatchNo");

    // Sayfalama standardı (T3.2): page 1'den başlar, pageSize üst sınırı 2000
    const page = Math.max(1, Number(searchParams.get("page")) || 1);
    const pageSize = Math.min(2000, Math.max(1, Number(searchParams.get("pageSize")) || 200));

    // Mağaza izolasyonu sunucuda, oturum claim'inden zorlanır (F-11):
    // istemciden gelen storeCode parametresi STORE_USER için yok sayılır.
    const effectiveStore = resolveStoreScope(currentUser, requestedStore);

    const conditions = [];
    if (effectiveStore && effectiveStore !== "ALL") {
      conditions.push(eq(orders.buyerStore, effectiveStore));
    }
    if (cargoStatus && cargoStatus !== "ALL") {
      conditions.push(eq(orders.cargoStatus, cargoStatus));
    }
    if (pshBatchNo && pshBatchNo !== "ALL") {
      conditions.push(eq(orders.pshBatchNo, pshBatchNo));
    }

    const ordersBase = db.select().from(orders);
    const allOrders = await (conditions.length > 0
      ? ordersBase
          .where(and(...conditions))
          .orderBy(desc(orders.orderDate), desc(orders.id))
          .limit(pageSize)
          .offset((page - 1) * pageSize)
      : ordersBase
          .orderBy(desc(orders.orderDate), desc(orders.id))
          .limit(pageSize)
          .offset((page - 1) * pageSize));

    // KPI seçimi SQL tarafında aggregate ile hesaplanir (T2.6):
    // bellekte tum tabloyu reduce etmek yok; sayfa boyutu buyurse bile sabit maliyet
    const kpiSelect = {
      totalOrdersCount: count(),
      totalUnits: sql<string>`coalesce(sum(${orders.quantity}), 0)`,
      totalSpend: sql<string>`coalesce(sum(${orders.totalCost}), 0)`,
      totalShippedToAmazon: sql<string>`coalesce(sum(${orders.shippedToAmazon}), 0)`,
      p1CancelTotal: sql<string>`coalesce(sum(${orders.p1CancelQty}), 0)`,
      p2MissingTotal: sql<string>`coalesce(sum(${orders.p2MissingQty}), 0)`,
      p3DefectiveTotal: sql<string>`coalesce(sum(${orders.p3DefectiveQty}), 0)`,
      p4ExpiredTotal: sql<string>`coalesce(sum(${orders.p4ExpiredQty}), 0)`,
      totalRefunds: sql<string>`coalesce(sum(${orders.refundAmount}), 0)`,
      problemOrdersCount: sql<string>`count(*) filter (where
        ${orders.cargoStatus} = 'İPTAL'
        or ${orders.p1CancelQty} > 0
        or ${orders.p2MissingQty} > 0
        or ${orders.p3DefectiveQty} > 0
        or ${orders.p4ExpiredQty} > 0
        or ${orders.refundAmount} > 0)`,
    };

    const kpiBase = db.select(kpiSelect).from(orders);
    const kpiQuery =
      conditions.length > 0 ? kpiBase.where(and(...conditions)) : kpiBase;

    const [[kpi], allStores, allBatches, allAuditLogs, allUsers] = await Promise.all([
      kpiQuery,
      db.select().from(stores).orderBy(stores.storeCode),
      effectiveStore && effectiveStore !== "ALL"
        ? db.select().from(pshBatches).where(eq(pshBatches.storeCode, effectiveStore)).orderBy(desc(pshBatches.createdAt))
        : db.select().from(pshBatches).orderBy(desc(pshBatches.createdAt)),
      db.select().from(auditLogs).orderBy(desc(auditLogs.createdAt)).limit(40),
      // password_hash asla istemciye taşınmaz — yalnızca güvenli alanlar seçilir
      db
        .select({
          id: users.id,
          name: users.name,
          email: users.email,
          role: users.role,
          storeCode: users.storeCode,
          avatar: users.avatar,
          createdAt: users.createdAt,
        })
        .from(users),
    ]);

    const totalOrdersCount = Number(kpi?.totalOrdersCount || 0);
    const totalUnits = Number(kpi?.totalUnits || 0);
    const totalSpend = Number(kpi?.totalSpend || 0);
    const totalShippedToAmazon = Number(kpi?.totalShippedToAmazon || 0);
    const p1CancelTotal = Number(kpi?.p1CancelTotal || 0);
    const p2MissingTotal = Number(kpi?.p2MissingTotal || 0);
    const p3DefectiveTotal = Number(kpi?.p3DefectiveTotal || 0);
    const p4ExpiredTotal = Number(kpi?.p4ExpiredTotal || 0);
    const totalRefunds = Number(kpi?.totalRefunds || 0);
    const problemOrdersCount = Number(kpi?.problemOrdersCount || 0);

    return NextResponse.json({
      orders: allOrders,
      stores: allStores,
      batches: allBatches,
      auditLogs: allAuditLogs,
      users: allUsers,
      // Anonim fallback kaldırıldı: oturum bu noktada garanti edilir (F-02/F-05)
      currentUser,
      pagination: {
        page,
        pageSize,
        total: totalOrdersCount,
        pageCount: Math.ceil(totalOrdersCount / pageSize),
      },
      kpis: {
        totalOrdersCount,
        totalUnits,
        totalSpend: totalSpend.toFixed(2),
        totalShippedToAmazon,
        problemOrdersCount,
        p1CancelTotal,
        p2MissingTotal,
        p3DefectiveTotal,
        p4ExpiredTotal,
        totalRefunds: totalRefunds.toFixed(2),
      },
    });
  } catch (error: unknown) {
    return handleRouteError("GET /api/orders", error);
  }
}

export async function POST(req: Request) {
  try {
    const gate = await requireUser();
    if (isDenied(gate)) return gate.response;
    const currentUser = gate.user;

    // Zod doğrulama (T3.1): 40 kolonluk sözleşme tek şemada merkezileşti
    const parsed = await parseBody(req, orderCreateSchema);
    if ("response" in parsed) return parsed.response;
    const body = parsed.data;

    // Mağaza kapsamı oturumdan zorlanır (F-11): STORE_USER başka mağazaya yazamaz
    const targetStore = resolveStoreScope(currentUser, body.buyerStore || "HRN");

    const {
      orderDate = new Date().toISOString().split("T")[0],
      imageUrl = "",
      fulfillmentType = "FBA",
      productTitle,
      asin,
      msku,
      supplierName = "THE VITAMINSHOPPE",
      supplierCode = "A198",
      supplierUrl = "",
      amazonUrl = "",
      orderNumber,
      driveLink = "",
      packCount = 1,
      quantity = 1,
      unitCost = 0,
      sellingPrice = 0,
      totalCost = 0,
      orderEmail = "cerberusnisan@gmail.com",
      cargoStatus = "Tam Geldi",
      shippedToAmazon = 0,
      p1CancelQty = 0,
      p2MissingQty = 0,
      p3DefectiveQty = 0,
      p4ExpiredQty = 0,
      problemAction = "",
      problemResult = "",
      refundAmount = 0,
      creditCard = "1753",
      isFragile = "NO",
      isMultiPack = "NO",
      isBundle = "NO",
      condition = "New",
      brandName = "General",
      description1 = "",
      description2 = "",
      auditNote = "",
      periodCode = "Ş26",
      correctedCost = 0,
      pshBatchNo = "",
      pshStatus = "BEKLIYOR",
      inventoryLabStatus = "GIRILMEDI",
    } = body;

    // Aktör adı istemciden değil oturumdan alınır (audit spoofing engeli)
    const actorName = currentUser.name;

    const calculatedTotal = Number(totalCost) || Number(unitCost) * Number(quantity);
    const calculatedCorrected = Number(correctedCost) || calculatedTotal;

    const [inserted] = await db
      .insert(orders)
      .values({
        buyerStore: targetStore,
        orderDate,
        imageUrl: imageUrl || "https://images.unsplash.com/photo-1584308666744-24d5c474f2ae?w=200&auto=format&fit=crop&q=80",
        fulfillmentType,
        productTitle,
        asin: asin.trim().toUpperCase(),
        msku: msku ? msku.trim() : `${targetStore}-${asin.trim()}`,
        supplierName,
        supplierCode,
        supplierUrl: supplierUrl || `https://www.vitaminshoppe.com/search?q=${encodeURIComponent(productTitle)}`,
        amazonUrl: amazonUrl || `https://www.amazon.com/dp/${asin.trim()}`,
        orderNumber: orderNumber.trim(),
        driveLink,
        packCount: Number(packCount) || 1,
        quantity: Number(quantity) || 1,
        unitCost: Number(unitCost).toFixed(2),
        sellingPrice: Number(sellingPrice).toFixed(2),
        totalCost: calculatedTotal.toFixed(2),
        orderEmail,
        cargoStatus,
        shippedToAmazon: Number(shippedToAmazon) || 0,
        p1CancelQty: Number(p1CancelQty) || 0,
        p2MissingQty: Number(p2MissingQty) || 0,
        p3DefectiveQty: Number(p3DefectiveQty) || 0,
        p4ExpiredQty: Number(p4ExpiredQty) || 0,
        problemAction,
        problemResult,
        refundAmount: Number(refundAmount).toFixed(2),
        creditCard,
        isFragile,
        isMultiPack,
        isBundle,
        condition,
        brandName,
        description1,
        description2,
        auditNote,
        periodCode,
        correctedCost: calculatedCorrected.toFixed(2),
        pshBatchNo: pshBatchNo || null,
        pshStatus,
        inventoryLabStatus,
      })
      .returning();

    // Audit log
    await db.insert(auditLogs).values({
      actorName,
      storeCode: targetStore,
      actionType: "ORDER_CREATED",
      targetEntity: `${orderNumber} - ${productTitle.slice(0, 32)}`,
      beforeState: "YENI_GIRIS",
      afterState: cargoStatus,
      details: `${targetStore} mağazasına ${quantity} adet (${unitCost}$) sipariş girildi.`,
    });

    return NextResponse.json({
      message: "Sipariş Google Drive XLS veritabanına başarıyla kaydedildi.",
      order: inserted,
    });
  } catch (error: unknown) {
    return handleRouteError("POST /api/orders", error);
  }
}